requirements.txt
asgiref==3.3.4
boto3==1.17.62
botocore==1.20.62
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-redis==4.12.1
django-simple-history==3.0.0
django-storages==1.11.1
django-taggit==1.3.0
django-tinymce4-lite==1.8.0
django-user-visit==0.4.1
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
jsmin==2.2.2
oauthlib==3.1.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.4
user-agents==2.2.0
